pyproject.toml
[tool.poetry]
name = "pythonProject5"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psycopg2 = "^2.9.9"
django = "^4.2.6"
pillow = "^10.1.0"
djangorestframework-simplejwt = "^5.3.0"
djangorestframework = "^3.14.0"
djoser = "^2.2.0"
drf-base64 = "^2.0"
drf-extra-fields = "^3.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
